# swagger/doc.json
{
	"openapi": "3.0.1",
	"info": { "title": "API https://tvoe-dev.ru/", "version": "v1" },
	"servers": [{ "url": "http://localhost:4000" }],
...
	"paths": {
		"/profile": {
			"get": {
				"summary": "Получение профиля",
				"tags": ["Profile"],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"_id": { "type": "string", "description": "Profile id" },
										"email": {
											"type": "string",
											"description": "Email user(profile)"
										},
										"avatar": {
											"type": "string",
											"description": "Path to img src profile user"
										},
										"firstname": {
											"type": "string",
											"description": "Имя этого человека"
										},
										"subscribe": {
											"type": "object",
											"description": "Информация по подписке. Если подписки нет, в этом поле null",
											"properties": {
												"startAt": {
													"type": "Date",
													"description": "Дата начала подписки"
												},
												"finishAt": {
													"type": "Date",
													"description": "Дата окончания подписки"
												},
												"tariffId": {
													"type": "string",
													"description": "Id тарифа (mongoose.ObjectId)"
												}
											}
										},
										"allowTrialTariff": {
											"type": "boolean",
											"description": "Позволен ли пробный тариф"
										},
										"role": {
											"type": "string",
											"description": "Admin, manager или, если это обычный пользователь - это поле отсуствует."
										}
									},
									"example": {
										"_id": "4180807e904d20e5d4c1b6qw",
										"email": "alex22@yandex.ru",
										"avatar": null,
										"firstname": "Иван",
										"allowTrialTariff": false,
										"role": "admin",
										"subscribe": {
											"startAt": "2021-04-13T07:41:42.000Z",
											"finishAt": "2021-04-13T07:41:42.000Z",
											"tariffId": "63dbe1687f457bc81bc920cb"
										}
									}
								}
							}
						}
					}
				}
			},
			"patch": {
				"summary": "Обновления профиля",
				"tags": ["Profile"],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"type": {
											"type": "string",
											"description": "Успешно или нет"
										},
										"alert": {
											"type": "boolean",
											"description": "Булевое значение"
										},
										"msg": {
											"type": "string",
											"description": "Сообщение \"Имя пользователя обновлено\""
										}
									},
									"example": {
										"type": "success",
										"alert": true,
										"description": "Имя пользователя обновлено"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"summary": "Удаление профиля",
				"tags": ["Profile"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"isRefund": {
										"type": "string",
										"description": "Желает ли пользователь вернуть средства средства за оставшийся период подписки."
									},
									"reason": {
										"type": "object",
										"description": "Причины удаления аккаунта",
										"properties": {
											"text": {
												"type": "string",
												"description": "Комментарий пользователя почему хочет удалить аккаунт"
											},
											"types": {
												"type": "array",
												"items": {
													"type": "string",
													"description": "Список выбранных пользователем причин - возможные значения - 'NOT_ENOUGH_CONTENT', 'BAD_QUALITY_VIDEO', 'BAD_SOUND', 'NOT_MATCH_TARIFF', 'HIGH_COST_SUBSCRIBTION'"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"deleted": {
											"type": "object",
											"description": "Информация по срокам удаления. Восстановить аккаунт можно только в течение одного месяца после удаления",
											"properties": {
												"start": {
													"type": "Date",
													"description": "Дата удаления"
												},
												"finish": {
													"type": "Date",
													"description": "Дата удаления + 1 месяц"
												}
											}
										}
									},
									"example": {
										"deleted": {
											"start": "2022-04-13T07:41:42.000Z",
											"finish": "2021-05-13T07:41:42.000Z"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/profile/devices": {
			"get": {
				"summary": "Профиль > Мои устройства",
				"tags": ["Profile"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "number" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "number" },
						"description": "Количество пропускаемых элементов из начала выборки"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"userToken": {
											"type": "string",
											"description": "JWT Токен юзера"
										},
										"totalSize": {
											"type": "number",
											"description": "Общее количество устройств, с которых был выполнен вход"
										},
										"sessions": {
											"type": "array",
											"description": "Список сессий",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"token": {
														"type": "string",
														"description": "JWT Токен юзера"
													},
													"ip": {
														"type": "string",
														"description": "Ip адресс юзера"
													},
													"os": {
														"type": "string",
														"description": "Операционная система устройства"
													},
													"isBot": {
														"type": "boolean",
														"description": "Является ли юзер ботом"
													},
													"isMobile": {
														"type": "boolean",
														"description": "Является ли дейвайс мобильным устройством"
													},
													"isDesktop": {
														"type": "boolean",
														"description": "Является ли дейвайс десктопным устройством"
													},
													"browser": {
														"type": "string",
														"description": "Название браузера"
													},
													"version": {
														"type": "string",
														"description": "Версия браузера"
													},
													"platform": {
														"type": "string",
														"description": "Платформа устройства ( ос)"
													},
													"createdAt": {
														"type": "Date",
														"description": "Дата создания"
													},
													"_id": {
														"type": "string",
														"description": "Id сессии (mongoose.Schema.Types.ObjectId)"
													}
												}
											}
										}
									},
									"example": {
										"userToken": "iyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0ODA4MDdlOTA0ZDIwZTVkNGMxYjZkYiIsImlhdCI6MTY4NjE1Mzk2OSwiZXhwIjoxNzE3NzExNTY5fQ.vY6dB6oVVF05afwr9-aD0oHZkELKRKS5YiJwvQs4A3M",
										"totalSize": 2,
										"sessions": [
											{
												"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0ODA4MDdlOTA0ZDIwZTVkNGMxYjZkYiIsImlhdCI6MTY4NjE1Mzk2OSwiZXhwIjoxNzE3NzExNTY5fQ.vY6dB6oVVF05afwr9-aD0oHZkELKRKS5YiJwvQs4A3M",
												"ip": "::ffff:127.0.0.1",
												"os": "Linux 64",
												"isBot": "false,",
												"isMobile": "false,",
												"isDesktop": "true,",
												"browser": "Chrome",
												"version": "103.0.0.0",
												"platform": "Linux",
												"createdAt": "2023-06-07T16:06:09.306Z",
												"_id": "6480aaf12fc38dca8042a3b5"
											},
											{
												"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0ODA4MDdlOTA0ZDIwZTVkNGMxYjZkYiIsImlhdCI6MTY4NjE1Mzk2OSwiZXhwIjoxNzE3NzExNTY5fQ.vY6dB6oVVF05afwr9-aD0oHZkELKRKS5YiJwvQs4A3M",
												"ip": "::ffff:127.0.0.1",
												"os": "Linux 64",
												"isBot": "false,",
												"isMobile": "false,",
												"isDesktop": "true,",
												"browser": "Firefox",
												"version": "113.0",
												"platform": "Linux",
												"createdAt": "2023-06-07T16:06:09.306Z",
												"_id": "6480aaf12fc38dca8042a3b5"
											}
										]
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/admin/users/profile": {
			"get": {
				"summary": "Админ-панель > Получение пользователя",
				"tags": ["Admin"],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"schema": { "type": "string" },
						"description": "id пользователя (mongoose.ObjectId)"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"user": {
											"type": "object",
											"description": "Общее количество сущностей в БД",
											"properties": {
												"_id": {
													"type": "string",
													"description": "Profile id"
												},
												"role": {
													"type": "string",
													"required": false,
													"description": "Роль этого профиля. Admin, manager или если это обычный пользователь то это поле отсуствует."
												},
												"email": {
													"type": "string",
													"description": "Email user(profile)"
												},
												"avatar": {
													"type": "string",
													"description": "Path to img src profile user"
												},
												"firstname": {
													"type": "string",
													"description": "Имя этого человека"
												},
												"deleted": {
													"type": "boolean",
													"description": "Удален ли аккаунт. Если не удален, этого поля нет",
													"required": false
												},
												"subscribe": {
													"type": "object",
													"description": "Информация по подписке. Если подписки нет, в этом поле null",
													"properties": {
														"startAt": {
															"type": "Date",
															"description": "Дата начала подписки"
														},
														"finishAt": {
															"type": "Date",
															"description": "Дата окончания подписки"
														},
														"tariffId": {
															"type": "string",
															"description": "Id тарифа (mongoose.ObjectId)"
														}
													}
												}
											}
										},
										"tariffs": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"_id": {
														"type": "string",
														"description": "Profile id"
													},
													"name": {
														"type": "string",
														"description": "Название тарифа"
													},
													"price": {
														"type": "number",
														"description": "Цена тарифа"
													},
													"sort": {
														"type": "string",
														"description": "Сорт тарифа?"
													},
													"hidden": {
														"type": "boolean",
														"description": "Скрыт ли тариф",
														"required": false
													}
												}
											}
										}
									}
								},
								"examples": {
									"example": {
										"value": {
											"user": {
												"email": "ovodenko.ksenia@yandex.ru",
												"avatar": null,
												"firstname": "ксения",
												"subscribe": {
													"startAt": "2023-02-12T04:39:28.501Z",
													"finishAt": "2023-05-13T04:39:28.501Z",
													"tariffId": "63dbe1447f457bc81bc920ca"
												}
											},
											"tariffs": [
												{
													"_id": "63f7537f841f04f2348f8de5",
													"name": "7 дней",
													"price": 0,
													"sort": 0
												},
												{
													"_id": "63dbda927f457bc81bc920c4",
													"name": "14 дней",
													"price": 0,
													"sort": 1,
													"hidden": true
												},
												{
													"_id": "63dbe11e7f457bc81bc920c9",
													"name": "1 месяц",
													"price": 199,
													"sort": 2
												},
												{
													"_id": "63dbe1447f457bc81bc920ca",
													"name": "3 месяца",
													"price": 499,
													"discount": 17,
													"sort": 3
												},
												{
													"_id": "63dbe1687f457bc81bc920cb",
													"name": "6 месяцев",
													"price": 899,
													"discount": 25,
													"sort": 4
												}
											]
										}
									}
								}
							}
						}
					}
				}
			},
			"patch": {
				"summary": "Админ-панель > Обновления пользователя",
				"tags": ["Admin"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"type": "string",
										"description": "Id пользователя (mongoose.ObjectId)"
									},
									"role": {
										"type": "string",
										"description": "Admin, manager или если это обычный пользователь то это поле отсуствует."
									},
									"tariffId": {
										"type": "string",
										"description": "Id тарифа (mongoose.ObjectId)"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"type": { "type": "string", "description": "success" },
										"msg": {
											"type": "string",
											"description": "Профиль обновлен"
										}
									},
									"example": { "type": "success", "msg": "Профиль обновлен" }
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}